+++ lib/lightning/usage_tracking/report.ex
@@ -13,13 +13,22 @@ defmodule Lightning.UsageTracking.Report do
     field :submitted, :boolean
     field :submitted_at, :utc_datetime_usec
     field :report_date, :date
+    field :submission_status, Ecto.Enum, values: [:pending, :success, :failure]
 
     timestamps()
   end
 
   def changeset(report, params) do
+    cast_params = [
+      :data,
+      :report_date,
+      :submission_status,
+      :submitted,
+      :submitted_at
+    ]
+
     report
-    |> cast(params, [:data, :report_date, :submitted, :submitted_at])
+    |> cast(params, cast_params)
     |> validate_required([:data, :report_date, :submitted])
     |> unique_constraint(:report_date)
   end