+++ priv/repo/migrations/20240402134814_add_submission_state_to_usage_tracking_reports.exs
@@ -1,0 +1,9 @@
+defmodule Lightning.Repo.Migrations.AddSubmissionStateToUsageTrackingReports do
+  use Ecto.Migration
+
+  def change do
+    alter table(:usage_tracking_reports) do
+      add :submission_status, :string, null: true
+    end
+  end
+end