+++ test/lightning/usage_tracking/report_worker_test.exs
@@ -72,7 +72,8 @@ defmodule Lightning.UsageTracking.ReportWorkerTest do
         %Report{
           submitted: true,
           report_date: ^report_date,
-          data: %{"instance" => ^expected_instance_data}
+          data: %{"instance" => ^expected_instance_data},
+          submission_status: :success
         } = report
       )
 
@@ -99,7 +100,8 @@ defmodule Lightning.UsageTracking.ReportWorkerTest do
           submitted: false,
           report_date: ^report_date,
           data: %{"instance" => ^expected_instance_data},
-          submitted_at: nil
+          submitted_at: nil,
+          submission_status: :failure
         } = report
       )
     end
@@ -170,7 +172,8 @@ defmodule Lightning.UsageTracking.ReportWorkerTest do
         %Report{
           submitted: true,
           report_date: ^report_date,
-          data: %{"instance" => ^expected_instance_data}
+          data: %{"instance" => ^expected_instance_data},
+          submission_status: :success
         } = report
       )
 
@@ -197,7 +200,8 @@ defmodule Lightning.UsageTracking.ReportWorkerTest do
           submitted: false,
           report_date: ^report_date,
           data: %{"instance" => ^expected_instance_data},
-          submitted_at: nil
+          submitted_at: nil,
+          submission_status: :failure
         } = report
       )
     end